pyproject.toml
[project]
name = "claptrap"
version = "0.1.0"
description = "A Discord bot for Borderlands fans."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.0.0",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",
    "anthropic>=0.34.0",
    "openai>=1.0.0",
    "youtube-transcript-api>=0.6.0",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.8.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.8.0",
    "trio>=0.30.0",
    "ty>=0.0.1a8",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "S", "C4", "PIE", "SIM", "PL"]
ignore = ["S101", "PLR0913", "PLR2004"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
